chore(STK-185): PR comment to rename Tab file in a meaningful way
