+++ packages/app/components/tabs/index.ts
@@ -1,1 +1,1 @@
-export * from "./constants";
+export * from "./styles";
